annotations: prevent generating .pyc files on import
We may end up adding these files to our git repos by mistake, so simply
prevent generating them in advance.

There's a tiny performance penalty with this change, because python
needs to re-generate the bytecode on-the-fly every time the annotations
script is called, but this overhead is absolutely negligible compared
the rest of the kernel build time.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>