annotations: update: update note when config is changing
When a new .config is imported some values may change, but the
corresponding note is not always updated to reflect the new value.

This is wrong and it can lead to errors and incorrect information in the
annotations file.

To prevent this from happening automatically add a TODO note when any
config with a note is changing to force the reviewer to update the note.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>